ZIGZAG
When a button is pressed, numbers from 1 to 200 will be written on the screen. However, if the number is a multiple of 3, 'zig' will be written instead of the number, and if the number is a multiple of 5, 'zag' will be written instead of the number. A zigzag will be written if the number is a multiple of both 3 and 5.